requirements.txt
torch                 2.5.1+cu121
openai                2.21.0
openai-whisper        20250625
sentence-transformers 5.2.3
openai
faiss-cpu             1.13.2
sounddevice           0.5.5